requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
motor==3.1.2
pymongo==4.3.3
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.28.2
python-multipart==0.0.6